AI-Resume-v1/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
rich==13.7.0
sqlmodel==0.0.14
pytest==8.3.3
pytest-cov==5.0.0
